+++ mayan/apps/converter/migrations/0001_initial.py
@@ -1,6 +1,6 @@
 from django.db import models, migrations
 
-import mayan.apps.converter.models
+import mayan.apps.common.validators
 
 
 class Migration(migrations.Migration):
@@ -38,7 +38,7 @@ class Migration(migrations.Migration):
                     'arguments', models.TextField(
                         blank=True, null=True, verbose_name='Arguments',
                         validators=[
-                            mayan.apps.converter.validators.YAMLValidator
+                            mayan.apps.common.validators.YAMLValidator
                         ]
                     )
                 ),